+++ src/swap/defi/cosmosIbc.ts
@@ -1,0 +1,104 @@
+import {
+  EdgeCorePluginOptions,
+  EdgeSpendInfo,
+  EdgeSwapInfo,
+  EdgeSwapPlugin,
+  EdgeSwapQuote,
+  EdgeSwapRequest,
+  SwapCurrencyError
+} from 'edge-core-js/types'
+
+import {
+  getMaxSwappable,
+  makeSwapPluginQuote,
+  SwapOrder
+} from '../../util/swapHelpers'
+import { convertRequest } from '../../util/utils'
+import { EdgeSwapRequestPlugin } from '../types'
+
+const swapInfo: EdgeSwapInfo = {
+  pluginId: 'cosmosibc',
+  displayName: 'Cosmos IBC',
+  orderUri: undefined,
+  supportEmail: 'support@edge.com'
+}
+
+export function makeCosmosIbcPlugin(
+  opts: EdgeCorePluginOptions
+): EdgeSwapPlugin {
+  const fetchSwapQuoteInner = async (
+    request: EdgeSwapRequestPlugin
+  ): Promise<SwapOrder> => {
+    const {
+      publicAddress: toAddress
+    } = await request.toWallet.getReceiveAddress({ tokenId: request.toTokenId })
+
+    const spendInfo: EdgeSpendInfo = {
+      tokenId: request.fromTokenId,
+      spendTargets: [
+        {
+          nativeAmount: request.nativeAmount,
+          publicAddress: toAddress
+        }
+      ],
+      assetAction: {
+        assetActionType: 'transfer'
+      },
+      swapData: {
+        isEstimate: false,
+        payoutAddress: toAddress,
+        plugin: { ...swapInfo },
+        payoutCurrencyCode: request.toCurrencyCode,
+        payoutNativeAmount: request.nativeAmount,
+        payoutWalletId: request.toWallet.id
+      },
+      savedAction: {
+        actionType: 'swap',
+        swapInfo,
+        isEstimate: false,
+        canBePartial: false,
+        fromAsset: {
+          pluginId: request.fromWallet.currencyInfo.pluginId,
+          tokenId: request.fromTokenId,
+          nativeAmount: request.nativeAmount
+        },
+        toAsset: {
+          pluginId: request.toWallet.currencyInfo.pluginId,
+          tokenId: request.toTokenId,
+          nativeAmount: request.nativeAmount
+        },
+        payoutAddress: toAddress,
+        payoutWalletId: request.toWallet.id
+      }
+    }
+
+    return {
+      request,
+      spendInfo,
+      swapInfo,
+      fromNativeAmount: request.nativeAmount
+    }
+  }
+
+  const out: EdgeSwapPlugin = {
+    swapInfo,
+
+    async fetchSwapQuote(req: EdgeSwapRequest): Promise<EdgeSwapQuote> {
+      const request = convertRequest(req)
+
+      if (
+        request.fromWallet.currencyInfo.pluginId ===
+          request.toWallet.currencyInfo.pluginId ||
+        request.fromCurrencyCode !== request.toCurrencyCode
+      ) {
+        throw new SwapCurrencyError(swapInfo, request)
+      }
+
+      const newRequest = await getMaxSwappable(fetchSwapQuoteInner, request)
+      const swapOrder = await fetchSwapQuoteInner(newRequest)
+      return await makeSwapPluginQuote(swapOrder)
+    }
+  }
+
+  return out
+}