+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Cosmos IBC transfer plugin
+
 ## 2.1.1 (2023-02-27)
 
 - added: (Swapuz) Above limit response handling